+++ grpcalchemy/globals.py
@@ -309,7 +309,7 @@ Working outside of application context.
 
 This typically means that you attempted to use functionality that needed
 to interface with the current application object in some way. To solve
-this, set up an application context with app.app_context().  See the
+this, set up an application context with app.app_context.  See the
 documentation for more information.\
 '''